+++ internal/database/migrations/migrations.go
@@ -59,6 +59,8 @@ var migrations = []Migration{
 	// on v22. Let's make a noop v22 to make sure every instance will not miss a
 	// real future migration.
 	NewMigration("noop", func(*gorm.DB) error { return nil }),
+	// v22 -> v23:v0.15.0
+	NewMigration("add commit_status.creator_name", addCommitStatusCreatorName),
 }
 
 var errMigrationSkipped = errors.New("the migration has been skipped")